cache pip packages
Adds improvements to run a faster workflow, and reduce overhead by caching the requirments. When a push or pull is done, and the workflow is triggered this should be able to grab a restore-key and use that. The cron job is set to run every sunday at midnight, clearing the cache, and creating a new one to ensure packages are up to date.